Move DistLearningConfig to config.py
