+++ fltk/util/config/__init__.py
@@ -6,8 +6,7 @@ import yaml
 import logging
 
 from fltk.util.config.distributed_config import DistributedConfig
-from fltk.util.config.config import Config, get_safe_loader
-from fltk.util.config.arguments import DistLearningConfig
+from fltk.util.config.config import Config, get_safe_loader, DistLearningConfig
 
 
 def retrieve_config_network_params(conf: Config, nic=None, host=None):
