+++ fltk/util/config/arguments.py
@@ -1,98 +1,7 @@
-import logging
 from argparse import ArgumentParser
-from dataclasses import dataclass
-from typing import List, Type, Dict, T, Union, Any
 
 import torch.distributed as dist
-import torch.nn
-from dataclasses_json import dataclass_json
 
-from fltk.util.config.definitions import Nets, Dataset
-
-_available_loss = {
-    "CROSSENTROPYLOSS": torch.nn.CrossEntropyLoss,
-    "HUBERLOSS" : torch.nn.HuberLoss
-}
-
-_available_optimizer: Dict[str, Type[torch.optim.Optimizer]] = {
-    "SGD": torch.optim.SGD,
-    "ADAM": torch.optim.Adam,
-    "ADAMW": torch.optim.AdamW
-}
-
-
-@dataclass_json
-@dataclass(frozen=True)
-class DistLearningConfig:  # pylint: disable=too-many-instance-attributes
-    """
-    Class encapsulating LearningParameters, for now used under DistributedLearning.
-    """
-    model: Nets
-    dataset: Dataset
-    batch_size: int
-    test_batch_size: int
-    max_epoch: int
-    learning_rate: float
-    learning_decay: float
-    loss: str
-    optimizer: str
-    optimizer_args: Dict[str, Any]
-    scheduler_step_size: int
-    scheduler_gamma: float
-    min_lr: float
-
-    cuda: bool
-    seed: int
-
-    @staticmethod
-    def __safe_get(lookup: Dict[str, T], keyword: str) -> T:
-        """
-        Static function to 'safe' get elements from a dictionary, to prevent issues with Capitalization in the code.
-        @param lookup: Lookup dictionary to 'safe get' from.
-        @type lookup: dict
-        @param keyword: Keyword to 'get' from the Lookup dictionary.
-        @type keyword: str
-        @return: Lookup value from 'safe get' request.
-        @rtype: T
-        """
-        safe_keyword = str.upper(keyword)
-        if safe_keyword not in lookup:
-            logging.fatal(f"Cannot find configuration parameter {keyword} in dictionary.")
-        return lookup.get(safe_keyword)
-
-    # def get_model_class(self) -> Type[torch.nn.Module]:
-    #     """
-    #     Function to obtain the model class that was given via commandline.
-    #     @return: Type corresponding to the model that was passed as argument.
-    #     @rtype: Type[torch.nn.Module]
-    #     """
-    #     return get_net(self.model)
-
-    # def get_dataset_class(self) -> Type[Dataset]:
-    #     """
-    #     Function to obtain the dataset class that was given via commandline.
-    #     @return: Type corresponding to the dataset that was passed as argument.
-    #     @rtype: Type[Dataset]
-    #     """
-    #     return get_dataset(self.dataset)
-
-    def get_loss(self) -> Type:
-        """
-        Function to obtain the loss function Type that was given via commandline to be used during the training
-        execution.
-        @return: Type corresponding to the loss function that was passed as argument.
-        @rtype: Type
-        """
-        return self.__safe_get(_available_loss, self.loss)
-
-    def get_optimizer(self) -> Type[torch.optim.Optimizer]:
-        """
-        Function to obtain the loss function Type that was given via commandline to be used during the training
-        execution.
-        @return: Type corresponding to the Optimizer to be used during training.
-        @rtype: Type[torch.optim.Optimizer]
-        """
-        return self.__safe_get(_available_optimizer, self.optimizer)
 
 def _create_extractor_parser(subparsers):
     """
